feat: add reader for GH issues
The reader retrieves issues from a repo a converts them into documents.
Documents's text is the concantenation of the issue title and body and the
document Id is the issue number. Further information about the issue is added to
the document extra_info field.

Issues can be filtered by their state (the default being open) and by label.