Update AzStorageBlobReader
Updated the reader to allow for utilization of connection string in replacement of an account URL and credentials. Account URL and credentials can still be used though.
Replaced the temporary download file path to use the random temp directory still, but the file uses the object name, in case the user needs that name during loading, and not the randomly generated name.
Updated README to reflect these changes, along with adding a tip on specifying their own file extractor, and how to add that to the required dictionary.
Updated library.json to add my username to this class.