Update AzStorageBlobReader
Added functionality to allow user to connect to blob storage with a connection string, instead of credentials and account URL. The credentials and URL
functionality still exists.
Updated README to reflect this new functionality, and added a tip on how to update file_extractor file to specify a specific different document reader, if needed.
Updated library to reflect myself as a contributor/author to this module.