Update AzStorageBlobReader (#756)
* Update AzStorageBlobReader

Updated the reader to allow for utilization of connection string in replacement of an account URL and credentials. Account URL and credentials can still be used though.
Replaced the temporary download file path to use the random temp directory still, but the file uses the object name, in case the user needs that name during loading, and not the randomly generated name.
Updated README to reflect these changes, along with adding a tip on specifying their own file extractor, and how to add that to the required dictionary.
Updated library.json to add my username to this class.

* Update AzStorageBlobReader

Added functionality to allow user to connect to blob storage with a connection string, instead of credentials and account URL. The credentials and URL
functionality still exists.
Updated README to reflect this new functionality, and added a tip on how to update file_extractor file to specify a specific different document reader, if needed.
Updated library to reflect myself as a contributor/author to this module.

* Lint

Fix linting issue.